feat: delete user as an admin

+++ src/application/user/user.service.ts
@@ -1,4 +1,4 @@
-import { Injectable, Logger } from '@nestjs/common'
+import { ConflictException, Injectable, Logger } from '@nestjs/common'
 
 import { IDatabaseConnection } from '../../persistence/database-connection.interface'
 
@@ -47,9 +47,14 @@ export class UserService implements IUserService {
     )
   }
 
-  public async deleteById(id: UserID): Promise<User> {
-    return this.databaseConnection.transactional(tx => {
-      return this.repository.deleteById(tx, id)
+  public async deleteById(id: UserID, currentUser: User): Promise<void> {
+    return this.databaseConnection.transactional(async tx => {
+      if (currentUser.id === id)
+        throw new ConflictException("You can't delete your own user account.")
+      await this.repository.deleteById(tx, id)
+      this.logger.verbose(
+        `User with id: ${id} has been deleted by admin: ${currentUser.id}`,
+      )
     })
   }
 }
